// libs/nestjs/common/enums.ts
export enum EnvEnum {
  STAGE = 'STAGE',
  DB_PORT = 'DB_PORT',
  DB_USERNAME = 'DB_USERNAME',
  USERNAME = 'USERNAME',
  DATABASE = 'DATABASE',
  SYNCRONIZE = 'SYNCRONIZE',
  JWT_SECRET_KEY = 'JWT_SECRET_KEY',
  JWT_SECRET = 'JWT_SECRET',
  PORT = 'PORT',
  HOST_API = 'HOST_API',
  API_URL = 'API_URL',
  DATABASE_URL = 'DATABASE_URL',
  DB_HOST = 'DB_HOST',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_NAME = 'DB_NAME',
  NODE_ENV = 'NODE_ENV',
  ALLOWED_ORIGINS = 'ALLOWED_ORIGINS',
  TZ = "TZ",
}


  